Moved `TimeSlot` out to it's own file
TimeSlot's use has outgrown just being part of free/busy. Now TimeSlot can be used in the availability endpoints so take it out and expand functionality.
